P4-3604 changing the container configuration to run as server class machine as recommended by cloud platform.
If a java/kotlin service pod has less than 2GB memory, without explicit configuration, it is most likely running a serial “stop everything” garbage collector, which has abysmal performance in production (pauses all threads while it’s cleaning)

Checking the production pods on the service proved this to be the case in that it is using the serial garbage collector.

Adding the flag the java run command addresses this.